Refactor into folders.
Better choose a folder organization now while there are not too many
files yet. I suspect the "widget" folder will need to be split again
when more custom widgets are added to the project.

The project file had to be moved to the root of the project tree,
because otherwise Qt behaves strangely when there is more than one
source folder.