Story 3: Can run truncated script.
Full story description:
I can run a prefix of the script using a command from the menu. I will
see the output of the truncated script on stdout, or an error if the
truncated code doesn't represent a valid Bash script.

The run action has a new name, "Run Entire Script", to distinguish it
from this new run action.