Begin iteration 2.
I picked the most important feature requests obtained from the previous
iteration, and converted them into stories. At this point of the
project, I believe it is more important to add features than to fix
bugs, because it is not yet clear if the features in which the bugs
manifest themselves will be kept in the long term.